Split SyntaxHighlighter.swift into model, walker, and custom parsers
SyntaxHighlighter.swift had grown to ~800 lines. Keep the public model
(Span/Kind) and the parse() entry point here, and move the implementation into
two documented siblings (pure code movement, no behavior change):

- SyntaxHighlighter+Walker.swift        — the swift-markdown AST walker
  (SpanCollector) and its source-offset helpers
- SyntaxHighlighter+CustomParsers.swift — regex passes for constructs the AST
  doesn't model (==highlight==, $math$, indented list items)

SpanCollector and the parser statics were widened from private to internal so
parse() can reach them across files.
